+++ pulsar-broker/src/main/java/org/apache/pulsar/broker/admin/v2/Worker.java
@@ -37,9 +37,7 @@ import java.util.function.Supplier;
 import javax.ws.rs.GET;
 import javax.ws.rs.Path;
 import javax.ws.rs.Produces;
-import javax.ws.rs.WebApplicationException;
 import javax.ws.rs.core.MediaType;
-import javax.ws.rs.core.Response;
 
 @Slf4j
 @Path("/worker")
@@ -116,22 +114,5 @@ public class Worker extends AdminResource implements Supplier<WorkerService> {
     @Produces(MediaType.APPLICATION_JSON)
     public List<ConnectorDefinition> getConnectorsList() throws IOException {
         return worker.getListOfConnectors(clientAppId());
-    }
-
-    @GET
-    @ApiOperation(
-            value = "Determines whether the worker service is initialized and ready for use",
-            response = Boolean.class
-    )
-    @ApiResponses(value = {
-            @ApiResponse(code = 400, message = "Invalid request"),
-            @ApiResponse(code = 408, message = "Request timeout")
-    })
-    @Path("/initialized")
-    public boolean isInitialized() {
-        if (!worker.isWorkerServiceInitialized()) {
-            throw new WebApplicationException(Response.Status.SERVICE_UNAVAILABLE);
-        }
-        return true;
     }
 }
