+++ pulsar-functions/worker/src/main/java/org/apache/pulsar/functions/worker/rest/Resources.java
@@ -18,10 +18,6 @@
  */
 package org.apache.pulsar.functions.worker.rest;
 
-import java.util.Arrays;
-import java.util.HashSet;
-import java.util.Set;
-
 import org.apache.pulsar.functions.worker.rest.api.FunctionsMetricsResource;
 import org.apache.pulsar.functions.worker.rest.api.v2.FunctionsApiV2Resource;
 import org.apache.pulsar.functions.worker.rest.api.v2.WorkerApiV2Resource;
@@ -32,6 +28,10 @@ import org.apache.pulsar.functions.worker.rest.api.v3.SinksApiV3Resource;
 import org.apache.pulsar.functions.worker.rest.api.v3.SourceApiV3Resource;
 import org.apache.pulsar.functions.worker.rest.api.v3.SourcesApiV3Resource;
 import org.glassfish.jersey.media.multipart.MultiPartFeature;
+
+import java.util.Arrays;
+import java.util.HashSet;
+import java.util.Set;
 
 public final class Resources {
 
@@ -64,7 +64,8 @@ public final class Resources {
         return new HashSet<>(
                 Arrays.asList(
                         ConfigurationResource.class,
-                        FunctionsMetricsResource.class
+                        FunctionsMetricsResource.class,
+                        WorkerReadinessResource.class
                 ));
     }
 }
