+++ pulsar-functions/worker/src/main/java/org/apache/pulsar/functions/worker/rest/api/v2/WorkerApiV2Resource.java
@@ -41,10 +41,8 @@ import javax.ws.rs.Consumes;
 import javax.ws.rs.GET;
 import javax.ws.rs.Path;
 import javax.ws.rs.Produces;
-import javax.ws.rs.WebApplicationException;
 import javax.ws.rs.core.Context;
 import javax.ws.rs.core.MediaType;
-import javax.ws.rs.core.Response;
 
 @Slf4j
 @Path("/worker")
@@ -139,22 +137,5 @@ public class WorkerApiV2Resource implements Supplier<WorkerService> {
     @Path("/connectors")
     public List<ConnectorDefinition> getConnectorsList() throws IOException {
         return worker.getListOfConnectors(clientAppId());
-    }
-
-    @GET
-    @ApiOperation(
-            value = "Determines whether the worker service is initialized and ready for use",
-            response = Boolean.class
-    )
-    @ApiResponses(value = {
-            @ApiResponse(code = 400, message = "Invalid request"),
-            @ApiResponse(code = 408, message = "Request timeout")
-    })
-    @Path("/initialized")
-    public boolean isInitialized() {
-        if (!worker.isWorkerServiceInitialized()) {
-            throw new WebApplicationException(Response.Status.SERVICE_UNAVAILABLE);
-        }
-        return true;
     }
 }
